move pubsub observer outside systems
