+++ core/src/world/services/redis_pubsub_events_observer.py
